[Ajustes Fase 2] Movendo lógica de listar pedidos para use case
